test: move tests to source to fix collectCoverageFrom
